+++ ControllerFramework/HighLevel/strategies/dev/mms.json
@@ -1,7 +1,5 @@
 [
-  { "action": -1, "start_point": [0, 100], "direction": "E" },
+  { "action": -1, "start_point": [0, 1000], "direction": "E" },
   { "action": "set_var", "name": "flag", "value": 1},
-  { "action": "if", "current_value": "flag", "compare_with": 1 },
-    { "action": "log", "content": "Flag is true" },
-  { "action": "endif" }
+  { "action": "drive", "point": [1000, 1000]}
 ]
